+++ jib-core/src/test/java/com/google/cloud/tools/jib/api/event/EventHandlersTest.java
@@ -57,11 +57,9 @@ public class EventHandlersTest {
     EventHandlers eventHandlers =
         EventHandlers.builder()
             .add(
-                new JibEventType<>(TestJibEvent1.class),
+                TestJibEvent1.class,
                 testJibEvent1 -> Assert.assertEquals("payload", testJibEvent1.getPayload()))
-            .add(
-                new JibEventType<>(TestJibEvent2.class),
-                testJibEvent2 -> testJibEvent2.sayHello("Jib"))
+            .add(TestJibEvent2.class, testJibEvent2 -> testJibEvent2.sayHello("Jib"))
             .add(jibEvent -> counter[0]++)
             .build();
     Assert.assertTrue(eventHandlers.getHandlers().containsKey(JibEvent.class));
@@ -93,15 +91,9 @@ public class EventHandlersTest {
 
     EventHandlers eventHandlers =
         EventHandlers.builder()
-            .add(
-                new JibEventType<>(TestJibEvent2.class),
-                testJibEvent2 -> emissions.add("handled 2 first"))
-            .add(
-                new JibEventType<>(TestJibEvent2.class),
-                testJibEvent2 -> emissions.add("handled 2 second"))
-            .add(
-                new JibEventType<>(TestJibEvent3.class),
-                testJibEvent3 -> emissions.add("handled 3"))
+            .add(TestJibEvent2.class, testJibEvent2 -> emissions.add("handled 2 first"))
+            .add(TestJibEvent2.class, testJibEvent2 -> emissions.add("handled 2 second"))
+            .add(TestJibEvent3.class, testJibEvent3 -> emissions.add("handled 3"))
             .add(jibEvent -> emissions.add("handled generic"))
             .build();