+++ jib-core/src/test/java/com/google/cloud/tools/jib/api/event/events/ProgressEventTest.java
@@ -18,7 +18,6 @@ package com.google.cloud.tools.jib.api.event.events;
 
 import com.google.cloud.tools.jib.api.event.BuildStepType;
 import com.google.cloud.tools.jib.api.event.EventHandlers;
-import com.google.cloud.tools.jib.api.event.JibEventType;
 import com.google.cloud.tools.jib.event.events.ProgressEvent;
 import com.google.cloud.tools.jib.event.progress.Allocation;
 import java.util.HashMap;
@@ -48,7 +47,7 @@ public class ProgressEventTest {
   }
 
   private static EventHandlers makeEventHandlers(Consumer<ProgressEvent> progressEventConsumer) {
-    return EventHandlers.builder().add(JibEventType.PROGRESS, progressEventConsumer).build();
+    return EventHandlers.builder().add(ProgressEvent.class, progressEventConsumer).build();
   }
 
   private static final double DOUBLE_ERROR_MARGIN = 1e-10;